Edition on Lecture 2
I notice that note is not the same as the one we used in class. I guess it is an older version. So I edited the notes based on this older version, and then added back some features you did in the newer version. I then added some proofs and explanations you mentioned in class. Hope this can help.