Edition of Lecture 3
Lecture 3 edited. I added details of some computations. One thing I am not sure: when proving FWL theorem the small "y" is used. But throughout the note (as well as the definition at the top) the big "Y" is used. I changed them back to big "Y" for consistency but I am not sure.